[WIP]feat: format input section
