+++ wdl-format/src/format.rs
@@ -3,19 +3,16 @@
 use std::fmt::Write;
 
 use anyhow::Result;
+use wdl_ast::v1::Decl;
 use wdl_ast::v1::DocumentItem;
+use wdl_ast::v1::Expr;
+use wdl_ast::v1::InputSection;
 use wdl_ast::v1::LiteralBoolean;
 use wdl_ast::v1::LiteralFloat;
 use wdl_ast::v1::LiteralInteger;
-use wdl_ast::v1::LiteralNull;
 use wdl_ast::v1::LiteralString;
-use wdl_ast::v1::MetadataArray;
-use wdl_ast::v1::MetadataObject;
-use wdl_ast::v1::MetadataObjectItem;
-use wdl_ast::v1::MetadataSection;
-use wdl_ast::v1::MetadataValue;
-use wdl_ast::v1::ParameterMetadataSection;
 use wdl_ast::v1::StringPart;
+use wdl_ast::v1::Type;
 use wdl_ast::v1::WorkflowDefinition;
 use wdl_ast::v1::WorkflowItem;
 use wdl_ast::AstNode;
@@ -34,6 +31,7 @@ use wdl_ast::VersionStatement;
 mod comments;
 mod format_state;
 mod import;
+mod metadata;
 
 use comments::format_inline_comment;
 use comments::format_preceding_comments;
@@ -48,18 +46,6 @@ pub trait Formattable {
     fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()>;
     /// Get the syntax element of the element.
     fn syntax_element(&self) -> SyntaxElement;
-}
-
-impl Formattable for Comment {
-    fn format(&self, buffer: &mut String, _state: &mut FormatState) -> Result<()> {
-        let comment = self.as_str().trim();
-        write!(buffer, "{}{}", comment, NEWLINE)?;
-        Ok(())
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        SyntaxElement::Token(self.syntax().clone())
-    }
 }
 
 impl Formattable for Version {
@@ -170,12 +156,6 @@ impl Formattable for Ident {
     }
 }
 
-// impl Ord for ImportStatement {
-//     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-//         self.uri().as_str().cmp(other.uri().as_str())
-//     }
-// }
-
 impl Formattable for LiteralBoolean {
     fn format(&self, buffer: &mut String, _state: &mut FormatState) -> Result<()> {
         match self.value() {
@@ -212,9 +192,9 @@ impl Formattable for LiteralInteger {
     }
 }
 
-impl Formattable for LiteralNull {
+impl Formattable for Type {
     fn format(&self, buffer: &mut String, _state: &mut FormatState) -> Result<()> {
-        buffer.push_str("null");
+        write!(buffer, "{}", self.syntax().to_string())?;
         Ok(())
     }
 
@@ -223,217 +203,93 @@ impl Formattable for LiteralNull {
     }
 }
 
-impl Formattable for MetadataObject {
+impl Formattable for Expr {
+    fn format(&self, buffer: &mut String, _state: &mut FormatState) -> Result<()> {
+        write!(buffer, "{}", self.syntax().to_string())?;
+        Ok(())
+    }
+
+    fn syntax_element(&self) -> SyntaxElement {
+        SyntaxElement::Node(self.syntax().clone())
+    }
+}
+
+impl Formattable for Decl {
     fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
         format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
 
-        let open_brace = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::OpenBrace)
-                .expect("Metadata Object should have an open brace")
-                .as_token()
-                .expect("Open brace should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&open_brace, buffer, state, true)?;
-        if state.interrupted() {
-            state.indent(buffer)?;
-            state.reset_interrupted();
-        }
-        buffer.push('{');
-        format_inline_comment(&open_brace, buffer, state, false)?;
-
-        state.increment_indent();
-
-        for item in self.items() {
-            // state.indent(buffer)?;
-            item.format(buffer, state)?;
-            buffer.push(',');
-            buffer.push_str(NEWLINE);
-        }
-
-        state.decrement_indent();
-
-        let close_brace = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::CloseBrace)
-                .expect("Metadata Object should have a close brace")
-                .as_token()
-                .expect("Close brace should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&close_brace, buffer, state, false)?;
+        let ty = self.ty();
         state.indent(buffer)?;
-        buffer.push('}');
-        format_inline_comment(&self.syntax_element(), buffer, state, true)?;
-
-        Ok(())
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        SyntaxElement::Node(self.syntax().clone())
-    }
-}
-
-impl Formattable for MetadataArray {
-    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
-        format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
-
-        let open_bracket = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::OpenBracket)
-                .expect("Metadata Array should have an open bracket")
-                .as_token()
-                .expect("Open bracket should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&open_bracket, buffer, state, true)?;
-        if state.interrupted() {
-            state.indent(buffer)?;
-            state.reset_interrupted();
-        }
-        buffer.push('[');
-        format_inline_comment(&open_bracket, buffer, state, false)?;
-
-        state.increment_indent();
-
-        for item in self.elements() {
-            state.indent(buffer)?;
-            item.format(buffer, state)?;
-            buffer.push(',');
-            buffer.push_str(NEWLINE);
-        }
-
-        state.decrement_indent();
-
-        let close_bracket = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::CloseBracket)
-                .expect("Metadata Array should have a close bracket")
-                .as_token()
-                .expect("Close bracket should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&close_bracket, buffer, state, false)?;
-        state.indent(buffer)?;
-        buffer.push(']');
-        format_inline_comment(&self.syntax_element(), buffer, state, true)?;
-
-        Ok(())
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        SyntaxElement::Node(self.syntax().clone())
-    }
-}
-
-impl Formattable for MetadataValue {
-    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
-        match self {
-            MetadataValue::String(s) => s.format(buffer, state),
-            MetadataValue::Boolean(b) => b.format(buffer, state),
-            MetadataValue::Float(f) => f.format(buffer, state),
-            MetadataValue::Integer(i) => i.format(buffer, state),
-            MetadataValue::Null(n) => n.format(buffer, state),
-            MetadataValue::Object(o) => o.format(buffer, state),
-            MetadataValue::Array(a) => a.format(buffer, state),
-        }
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        match self {
-            MetadataValue::String(s) => s.syntax_element(),
-            MetadataValue::Object(o) => o.syntax_element(),
-            MetadataValue::Array(a) => a.syntax_element(),
-            MetadataValue::Boolean(b) => b.syntax_element(),
-            MetadataValue::Float(f) => f.syntax_element(),
-            MetadataValue::Integer(i) => i.syntax_element(),
-            MetadataValue::Null(n) => n.syntax_element(),
-        }
-    }
-}
-
-impl Formattable for MetadataObjectItem {
-    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
-        format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
+        ty.format(buffer, state)?;
+        format_inline_comment(&ty.syntax_element(), buffer, state, true)?;
 
         let name = self.name();
-        state.indent(buffer)?;
+        format_preceding_comments(&name.syntax_element(), buffer, state, true)?;
+        state.space_or_indent(buffer)?;
         name.format(buffer, state)?;
         format_inline_comment(&name.syntax_element(), buffer, state, true)?;
 
-        let colon = SyntaxElement::Token(
+        if let Some(expr) = self.expr() {
+            let eq = SyntaxElement::Token(
+                self.syntax()
+                    .children_with_tokens()
+                    .find(|element| element.kind() == SyntaxKind::Assignment)
+                    .expect("Bound declaration should have an equals sign")
+                    .as_token()
+                    .expect("Equals sign should be a token")
+                    .clone(),
+            );
+            format_preceding_comments(&eq, buffer, state, true)?;
+            state.space_or_indent(buffer)?;
+            buffer.push('=');
+            format_inline_comment(&eq, buffer, state, true)?;
+
+            format_preceding_comments(&expr.syntax_element(), buffer, state, true)?;
+            state.space_or_indent(buffer)?;
+            expr.format(buffer, state)?;
+        }
+        format_inline_comment(&self.syntax_element(), buffer, state, false)?;
+
+        Ok(())
+    }
+
+    fn syntax_element(&self) -> SyntaxElement {
+        SyntaxElement::Node(self.syntax().clone())
+    }
+}
+
+impl Formattable for InputSection {
+    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
+        format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
+
+        let input_keyword = SyntaxElement::Token(
             self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::Colon)
-                .expect("Metadata Object Item should have a colon")
-                .as_token()
-                .expect("Colon should be a token")
+                .first_token()
+                .expect("Input Section should have a token")
                 .clone(),
         );
-        format_preceding_comments(&colon, buffer, state, true)?;
-        if state.interrupted() {
-            state.indent(buffer)?;
-            state.reset_interrupted();
-        }
-        buffer.push(':');
-        format_inline_comment(&colon, buffer, state, true)?;
+        state.indent(buffer)?;
+        buffer.push_str("input");
+        format_inline_comment(&input_keyword, buffer, state, true)?;
 
-        let value = self.value();
-        format_preceding_comments(&value.syntax_element(), buffer, state, true)?;
+        let open_brace = SyntaxElement::Token(
+            self.syntax()
+                .children_with_tokens()
+                .find(|element| element.kind() == SyntaxKind::OpenBrace)
+                .expect("Input Section should have an open brace")
+                .as_token()
+                .expect("Open brace should be a token")
+                .clone(),
+        );
+        format_preceding_comments(&open_brace, buffer, state, true)?;    
         state.space_or_indent(buffer)?;
-        value.format(buffer, state)?;
-        format_inline_comment(&self.syntax_element(), buffer, state, false)?;
-
-        Ok(())
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        SyntaxElement::Node(self.syntax().clone())
-    }
-}
-
-impl Formattable for MetadataSection {
-    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
-        format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
-
-        let meta_keyword = SyntaxElement::Token(
-            self.syntax()
-                .first_token()
-                .expect("Metadata Section should have a token")
-                .clone(),
-        );
-        state.indent(buffer)?;
-        buffer.push_str("meta");
-        format_inline_comment(&meta_keyword, buffer, state, true)?;
-
-        let open_brace = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::OpenBrace)
-                .expect("Metadata Section should have an open brace")
-                .as_token()
-                .expect("Open brace should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&open_brace, buffer, state, true)?;
-        if !state.interrupted() {
-            buffer.push(' ');
-        } else {
-            state.reset_interrupted();
-            state.indent(buffer)?;
-        }
         buffer.push('{');
         format_inline_comment(&open_brace, buffer, state, false)?;
 
         state.increment_indent();
 
-        for item in self.items() {
-            item.format(buffer, state)?;
+        for decl in self.declarations() {
+            decl.format(buffer, state)?;
         }
 
         state.decrement_indent();
@@ -442,70 +298,7 @@ impl Formattable for MetadataSection {
             self.syntax()
                 .children_with_tokens()
                 .find(|element| element.kind() == SyntaxKind::CloseBrace)
-                .expect("Metadata Section should have a close brace")
-                .as_token()
-                .expect("Close brace should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&close_brace, buffer, state, false)?;
-        state.indent(buffer)?;
-        buffer.push('}');
-        format_inline_comment(&self.syntax_element(), buffer, state, false)?;
-
-        Ok(())
-    }
-
-    fn syntax_element(&self) -> SyntaxElement {
-        SyntaxElement::Node(self.syntax().clone())
-    }
-}
-
-impl Formattable for ParameterMetadataSection {
-    fn format(&self, buffer: &mut String, state: &mut FormatState) -> Result<()> {
-        format_preceding_comments(&self.syntax_element(), buffer, state, false)?;
-
-        let parameter_meta_keyword = SyntaxElement::Token(
-            self.syntax()
-                .first_token()
-                .expect("Parameter Metadata Section should have a token")
-                .clone(),
-        );
-        state.indent(buffer)?;
-        buffer.push_str("parameter_meta");
-        format_inline_comment(&parameter_meta_keyword, buffer, state, true)?;
-
-        let open_brace = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::OpenBrace)
-                .expect("Parameter Metadata Section should have an open brace")
-                .as_token()
-                .expect("Open brace should be a token")
-                .clone(),
-        );
-        format_preceding_comments(&open_brace, buffer, state, true)?;
-        if !state.interrupted() {
-            buffer.push(' ');
-        } else {
-            state.reset_interrupted();
-            state.indent(buffer)?;
-        }
-        buffer.push('{');
-        format_inline_comment(&open_brace, buffer, state, false)?;
-
-        state.increment_indent();
-
-        for item in self.items() {
-            item.format(buffer, state)?;
-        }
-
-        state.decrement_indent();
-
-        let close_brace = SyntaxElement::Token(
-            self.syntax()
-                .children_with_tokens()
-                .find(|element| element.kind() == SyntaxKind::CloseBrace)
-                .expect("Parameter Metadata Section should have a close brace")
+                .expect("Input Section should have a close brace")
                 .as_token()
                 .expect("Close brace should be a token")
                 .clone(),
@@ -577,7 +370,7 @@ impl Formattable for WorkflowDefinition {
                     pm.format(&mut parameter_meta_section_str, state)?;
                 }
                 WorkflowItem::Input(i) => {
-                    // i.format(&mut input_section_str, state)?;
+                    i.format(&mut input_section_str, state)?;
                 }
                 WorkflowItem::Call(c) => {
                     // c.format(&mut body_str, state)?;
@@ -605,6 +398,9 @@ impl Formattable for WorkflowDefinition {
         }
         if !parameter_meta_section_str.is_empty() {
             buffer.push_str(&parameter_meta_section_str);
+        }
+        if !input_section_str.is_empty() {
+            buffer.push_str(&input_section_str);
         }
 
         state.decrement_indent();
@@ -702,7 +498,6 @@ pub fn format_document(code: &str) -> Result<String, Vec<Diagnostic>> {
     }
 
     for item in ast.items() {
-        result.push_str(NEWLINE);
         match item.format(&mut result, &mut state) {
             Ok(_) => {}
             Err(_) => {
