+++ wdl-format/tests/format/every_token_no_comments/source.formatted
@@ -1,5 +1,4 @@
 version 1.0
-
 workflow test_wf {
     meta {
         a: "hello"
@@ -16,57 +15,47 @@ workflow test_wf {
                 1,
                 2,
                 3,
-            ]
-,
+            ],
             b: [
                 "hello",
                 "world",
                 "!",
-            ]
-,
+            ],
             c: {
-                x: 1
-,
-                y: 2
-,
-                z: 3
-,
-            }
-,
+                x: 1,
+                y: 2,
+                z: 3,
+            },
         }
         k: [
             {
                 a: {
-                }
-,
-                b: 0
-,
-                c: ""
-,
-                d: ""
-,
+                },
+                b: 0,
+                c: "",
+                d: "",
                 e: [
-                ]
-,
+                ],
             },
             {
                 x: [
                     1.0,
                     2.0,
                     3.0,
-                ]
-,
+                ],
             },
         ]
     }
     parameter_meta {
         out_sj_filter_overhang_min: {
-            type: "SpliceJunctionMotifs"
-,
-            label: "Minimum overhang required to support a splicing junction"
-,
+            type: "SpliceJunctionMotifs",
+            label: "Minimum overhang required to support a splicing junction",
         }
     }
+    input {
+        SpliceJunctionMotifs out_sj_filter_overhang_min = SpliceJunctionMotifs {
+noncanonical_motifs: 30,
+GT_AG_and_CT_AC_motif: 12,
 }
-
-
+    }
+}
