+++ wdl-format/tests/format/imports_with_both_comments/source.formatted
@@ -21,9 +21,5 @@ import  # fileA 1.2
 import "fileB.wdl" as foo  # also fileB
 # this comment belongs to fileC
 import "fileC.wdl"  # also fileC
-
-
-
 workflow test {
 }
-
