+++ wdl-format/tests/format/imports_with_inline_comments/source.formatted
@@ -10,9 +10,5 @@ import  # fileA 1
     Qux  # fileA 8
 import "fileB.wdl" as foo  # fileB
 import "fileC.wdl"  # fileC
-
-
 workflow test {
 }
-
-
