+++ wdl-format/tests/format/imports_with_no_comments/source.formatted
@@ -3,9 +3,5 @@ version 1.1
 import "fileA.wdl" as bar alias cows as horses alias cats as dogs
 import "fileB.wdl" as foo
 import "fileC.wdl" alias qux as Qux
-
-
-
 workflow test {
 }
-
